requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
elasticsearch==8.12.0
elasticsearch-dsl==8.12.0
psycopg2-binary==2.9.9
django-filter==24.1
Faker==22.5.1
